requirements.txt
# core
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
sqlalchemy
typing_extensions
python-dotenv

# DB driver (use ONE of these)
pymysql
# mysql-connector-python
